fix: reverted to long type for entityIDs

+++ Mooc.Application.Contracts/System/IUserService.cs
@@ -4,12 +4,12 @@ namespace Mooc.Application.Contracts.System
 {
     public interface IUserService
     {
-        Task<UserOutputDto> GetByIdAsync(int id);
+        Task<UserOutputDto> GetByIdAsync(long id);
         Task<UserOutputDto> GetByUserNameAsync(string userName);
         Task<UserListOutputDto> GetListAsync(int page, int pageSize);
-        Task<int> CreateAsync(CreateUserDto input, string hashedPassword);
-        Task DeleteAsync(List<int> ids);
-        Task UpdateAsync(int id, UpdateUserDto input, string? hashedPassword);
-        Task<UserRoleMetaDto> GetRoleMetaAsync(int id);
+        Task<long> CreateAsync(CreateUserDto input, string hashedPassword);
+        Task DeleteAsync(List<long> ids);
+        Task UpdateAsync(long id, UpdateUserDto input, string? hashedPassword);
+        Task<UserRoleMetaDto> GetRoleMetaAsync(long id);
     }
 }
